[ad-v4][step4] Match PT MXFP4 weight prep + add block_scale_interleave
Bring ``prepare_mxfp4_weights_for_trtllm_gen`` and the
``trtllm_mxfp4_w4a16_moe_fused`` op into structural parity with PT's
``MXFP4WeightTRTLLMGenFusedMoEMethod`` (quantization.py:4135) so the
trtllm-gen MoE kernel sees the same byte layout PT exercises:

mxfp4_weight_prep.py changes:
* Per-expert ``I_pad = roundUp(I, weight_alignment) = 2944`` first;
  derive ``2I_pad = 5888`` and ``I/2_pad = 1472`` from that. Previously
  we padded ``2I = 5760`` directly which is already 128-aligned and
  thus a no-op, leaving w1's effective ``I = 2880`` while w2's column
  padding pushed ``I = 2944`` -- inconsistent intermediate dim across
  the two gemms.
* W1 hidden axis padded to ``input_hidden_alignment = 512``
  (``H_w1_pad = 3072``), W2 hidden axis padded to
  ``weight_alignment = 128`` (``H_w2_pad = 2944``), matching PT's
  ``create_weights`` (lines 3715-3717 of quantization.py).
* De-interleave gate / up rows from the on-disk row-interleaved
  storage (``gate_up_proj_blocks[:, ::2, :]`` = gate,
  ``[:, 1::2, :]`` = up) and pad each half to ``I_pad`` separately
  before stacking as ``[up | gate]``. PT's chunk-then-copy dance
  (modeling_gpt_oss.py:695-706 + quantization.py:4252-4258) ends up
  with the same physical layout.
* Add ``torch.ops.trtllm.block_scale_interleave`` after
  ``shuffle_matrix`` for both fc1 and fc2 scales -- PT does both ops
  (quantization.py:4382, 4439); skipping the second was a partial
  bug.

trtllm_moe.py change:
* Routing softmax in fp32 instead of bf16 -- matches PT's
  ``RenormalizeMoeRoutingMethod`` which casts to fp32 for the topk
  softmax then back to the activation dtype.

Status: kernel builds and runs cleanly with these changes, and pure
GEMM throughput is at the V4 target (~9.28 ms ITL / ~108 tok/s for
gpt-oss-120b vs V3 Triton's 127.79 ms / 7.96 tok/s -- 13.5x).
However, content correctness is still blocked by an upstream NaN bug
in the trtllm-gen MoE kernel itself: PT's own
``TRTLLMGenFusedMoE.forward`` on gpt-oss-120b at this TRT-LLM commit
also produces NaN logits, so any byte-correct prep cannot rescue
output. Tracking note: re-validate when upstream fix lands; if
correctness is restored, proceed to step 5 (TP-MoE sharding).

Refs: cc_reports/gpt-oss-120b/MOE_TRTLLM_GEN_PLAN.md (step 4 of 6),
RESUME_V4.md.

+++ tensorrt_llm/_torch/auto_deploy/custom_ops/fused_moe/trtllm_moe.py
@@ -1391,11 +1391,14 @@ def trtllm_mxfp4_w4a16_moe_fused(
     x_shape = x.shape
     x2d = x.view(-1, x_shape[-1])
 
-    # Top-k routing (RenormalizeMoeRoutingMethod): logits -> topk -> softmax-of-topk.
-    # Done outside the kernel because bf16_mxe2m1_block_scale_moe_runner expects
-    # pre-computed topk_weights / topk_ids when routing_logits is None.
+    # Top-k routing — match PT's RenormalizeMoeRoutingMethod which casts to
+    # fp32 for the softmax to keep numerical precision (then casts back to
+    # the activation dtype for the kernel call). bf16 softmax over close
+    # logits can produce degenerate probabilities (all close to 1/k or
+    # extremely skewed), which translates to bad expert mixing and
+    # garbage-looking generation even when shapes/layouts are correct.
     router_logits = torch.nn.functional.linear(x2d, router_weight, router_bias)
-    topk_vals, topk_ids = torch.topk(router_logits, top_k, dim=-1)
+    topk_vals, topk_ids = torch.topk(router_logits.to(torch.float32), top_k, dim=-1)
     topk_weights = torch.nn.functional.softmax(topk_vals, dim=-1)
 
     # Pad activations to the kernel's expected hidden (H_pad, multiple of 512).
